+++ casanovo/denovo/model_runner.py
@@ -42,6 +42,7 @@ class ModelRunner:
         self,
         config: Config,
         model_filename: Optional[str] = None,
+        root_checkpoint_name: Optional[str] = None,
     ) -> None:
         """Initialize a ModelRunner"""
         self.config = config
@@ -54,6 +55,10 @@ class ModelRunner:
         self.loaders = None
         self.writer = None
 
+        checkpoint_filename = None
+        if root_checkpoint_name is not None:
+            checkpoint_filename = root_checkpoint_name + ".{epoch}-{step}"
+
         # Configure checkpoints.
         if config.save_top_k is not None:
             self.callbacks = [
@@ -62,6 +67,7 @@ class ModelRunner:
                     monitor="valid_CELoss",
                     mode="min",
                     save_top_k=config.save_top_k,
+                    filename=checkpoint_filename,
                 )
             ]
         else:
@@ -146,7 +152,9 @@ class ModelRunner:
         -------
         self
         """
-        self.writer = ms_io.MztabWriter(Path(output).with_suffix(".mztab"))
+        self.writer = ms_io.MztabWriter(
+            output.with_suffix(output.suffix + ".mztab")
+        )
         self.writer.set_metadata(
             self.config,
             model=str(self.model_filename),